vulnix module: build store derivations for system on the node
The store derivations that vulnix needs to scan the current system (with `--system`) had to be present on the node.
This change removes that requirement by building them on the node.

To be precise, the flake's nixosConfiguration matching the name of the node is built and we assume that to be the config that is actually deployed.