Fix issue with unsupported instance types
in a given region.

When launching instances without explicitly specifying an availability
zone, AWS picks one for us. However, it is not smart enough to verify
that said availability zone actually supports the requested instance
type. This notably affects us-east-1, but apparently not the
region/type combos we use in most our bitte deployment (which would
explain why we haven't observed this so far).

To solve this I've created a terraform module that leverages the
`aws_ec2_instance_type_offerings` data source and the result into a more
readily usable form (see
modules/terraform/modules/instance-types-to-azs/README.md) which is then
used to explicitly define an availability zone where necessary.

Note that this solution was "forward-ported" from an old branch that
has never been merged. For that reason it also pulls in some unrelated
code that implements CIDR calculations instead of hardcoding them. It
was intermingled with the needed code so I figured that spending
additional time separating things would be of
little to no value...