Use the instance_types_to_azs module for clients
as well. This mitigates an issue previously observed in core nodes,
where in some regions (notably us-east-1), when an availability zone is
not explicitly provided, AWS arbitrarily picks a zone which does not
necessarily support the requested instance types. This results in
intermittent an intermittent error that's not easily resolvable other
than destroy/plan/apply ad nauseam.