Factor `while` out of ssh invokation
it seems as though we still get a connection refused there, in which
case the script simply moves on (due to no `set -e`) and that's how we
end up with an empty `vaule.enc.json` in our local `encrypted` root.
this can subsequentally get copied to future deployments and cause much hairpull.