Add `core-` prefix to clients iam role name
this prevents a name conflict with the current role and thereby
contributes to a more seamless migration where we first apply core,
which creates the new role with the slightly different name, and then we
can apply clients, which would hook everything up to the new role and
delete the old one. This, as opposed to: applying clients first, to
remove the role from there, at which point the newly introduced `data`
source which retrieves the role attributes will have potentially errored
out (I have not tested that hypothesis but I believe it's plausible...),
then applying core, to create the (now non-conflicting) iam role, then
reapplying clients to recover.