consolidate vault policies into hydration
Before there was a conflict between the policies set on the core
machines at activation time, and the policies set by hydrate-cluster.
Specifically, the routing policy would always be deleted by terraform.

With this, we can ensure that terraform is aware of all of our defaut
vault policies so rebuilding and/or hydrating doesn't unexpectedly
change them.