bitte-cli: make devshell actually useful
Move the package and source under the expected `std` path and replace
it with a symlink so that I can reference it from the automation cell
allowing me to wire up the `RUST_SRC_PATH` so rust-anaylzer works as
expected and the rust toolchain is available.

Also move `.envrc` under this directory since it is only useful when
working on the cli.