add concept of private flake inputs
Inputs can now be declared without propagating to a consumers lock file
by adding inputs to the flake in the private sub directory.

The inputs defined here can be managed just like other flake inputs.

The narHash in the main flake.nix for the private inputs must be
updated anytime the private flake is touched.